+++ rotkehlchen/db/schema.py
@@ -56,6 +56,8 @@ INSERT OR IGNORE INTO location(location, seq) VALUES ('P', 16);
 INSERT OR IGNORE INTO location(location, seq) VALUES ('Q', 17);
 /* Bitstamp */
 INSERT OR IGNORE INTO location(location, seq) VALUES ('R', 18);
+/* Binance US */
+INSERT OR IGNORE INTO location(location, seq) VALUES ('S', 19);
 """
 
 # Custom enum table for AssetMovement categories (deposit/withdrawal)