+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 * :bug:`1969` Users who were using open nodes only and were seeing an out of gas error during defi balances query, should be able to query defi balances properly again.
 * :bug:`1287` Querying bitmex balances should now work properly again.
+* :feature:`1515` Rotki now supports Binance US. Users can see their balances and import trades, deposits and withdrawals from that exchange. They are also taken into account in the tax report.
 * :bug:`1916` Querying bitstamp trades should now work properly again.
 * :bug:`1917` Users can now properly login if they input the username after the password.
 * :bug:`1953` Show a proper error when a user inputs an invalid xpub or derivation path.